requirements.txt
asgiref==3.5.2
asttokens==2.4.0
backcall==0.2.0
black==22.10.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
comm==0.1.4
contourpy==1.1.0
coreapi==2.3.3
coreschema==0.0.4
crispy-bootstrap5==0.7
cycler==0.11.0
debugpy==1.6.7.post1
decorator==5.1.1
Django==4.1.3
django-bootstrap-daterangepicker==1.1.0
django-crispy-forms==1.14.0
django-dynamic-breadcrumbs==0.4.2
django-environ==0.9.0
django-view-breadcrumbs==2.4.1
djangorestframework==3.14.0
drf-yasg==1.21.4
executing==1.2.0
flake8==5.0.4
fonttools==4.42.1
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
ipykernel==6.25.2
ipython==8.15.0
itypes==1.2.0
jedi==0.19.0
Jinja2==3.1.2
jupyter_client==8.3.1
jupyter_core==5.3.1
kiwisolver==1.4.5
MarkupSafe==2.1.1
matplotlib==3.7.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mypy-extensions==0.4.3
nanoid==2.0.0
nest-asyncio==1.5.7
numpy==1.25.2
packaging==21.3
pandas==2.1.0
parso==0.8.3
pathspec==0.10.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.5.4
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.9.1
pyflakes==2.5.0
Pygments==2.16.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
pyzmq==25.1.1
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
seaborn==0.12.2
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
tomli==2.0.1
tornado==6.3.3
traitlets==5.9.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.12
wcwidth==0.2.6
